requirements.txt
awscli==1.11.69
awsebcli==3.10.1
blessed==1.14.2
boto3==1.4.4
botocore==1.5.32
cement==2.8.2
colorama==0.3.7
Django==1.10.6
django-cors-headers==2.0.2
django-extensions==1.7.8
django-rest-auth==0.9.1
django-storages==1.5.2
djangorestframework==3.6.2
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.13.1
jmespath==0.9.2
olefile==0.44
pathspec==0.5.0
Pillow==4.0.0
psycopg2==2.7.1
pyasn1==0.2.3
python-dateutil==2.6.0
PyYAML==3.12
requests==2.13.0
rsa==3.4.2
s3transfer==0.1.10
semantic-version==2.5.0
six==1.10.0
tabulate==0.7.5
termcolor==1.1.0
wcwidth==0.1.7
websocket-client==0.40.0
